+++ types/es6-shim/es6-shim-tests.ts
@@ -1,23 +1,18 @@
-declare const require: (module: string) => Object;
-if (require !== null) {
-  require('es6-shim');
-}
-
 interface Point { x: number; y: number; }
 interface Point3D extends Point { z: number; }
 
 let a: any;
-let s: string = '';
-let i: number = 2;
+let s = '';
+let i = 2;
 let iOrUndef: number | undefined;
 let b: boolean;
 let f: () => void = () => {};
 let o: Object;
 let r: RegExp = /a/;
-let sym: symbol = {} as symbol;
+declare const sym: symbol;
 let e: Error = new Error();
 let date: Date;
-let key: KeyOfProperty;
+let key: PropertyKey;
 let point: Point = { x: 1, y: 2 };
 let point3d: Point3D = { x: 1, y: 2, z: 3 };
 let point3dOrUndef: Point3D | undefined;
@@ -25,9 +20,9 @@ let pointOrUndef: Point | undefined;
 let arrayOfPoint: Point[] = [];
 let arrayOfPoint3D: Point3D[];
 let arrayOfSymbol: symbol[];
-let arrayOfPropertyKey: KeyOfProperty[];
+let arrayOfPropertyKey: PropertyKey[];
 let arrayOfAny: any[];
-let arrayOfStringAny: [string, any][];
+let arrayOfStringAny: Array<[string, any]>;
 let arrayLikeOfAny: ArrayLike<any> = [];
 let iterableOfPoint: IterableShim<Point> = [];
 let iterableOfStringPoint: IterableShim<[string, Point]> = [];
@@ -40,8 +35,8 @@ let iterableIteratorOfPointPoint: IterableIteratorShim<[Point, Point]>;
 let iterableIteratorOfNode: IterableIteratorShim<Node>;
 let iterableIteratorOfStringPoint: IterableIteratorShim<[string, Point]>;
 let iterableIteratorOfAny: IterableIteratorShim<any>;
-let iterableIteratorOfPropertyKey: IterableIteratorShim<KeyOfProperty>;
-let iterableIteratorOfPropertyKeyPoint: IterableIteratorShim<[KeyOfProperty, Point]>;
+let iterableIteratorOfPropertyKey: IterableIteratorShim<PropertyKey>;
+let iterableIteratorOfPropertyKeyPoint: IterableIteratorShim<[PropertyKey, Point]>;
 let nodeList: NodeList;
 let pd: PropertyDescriptor = {};
 let pdm: PropertyDescriptorMap = {};
@@ -86,7 +81,7 @@ b = Number.isSafeInteger(i);
 i = Number.MAX_SAFE_INTEGER;
 i = Number.MIN_SAFE_INTEGER;
 i = Number.parseFloat(s);
-i = Number.parseInt(s);
+i = Number.parseInt(s, 10);
 i = Number.parseInt(s, i);
 i = Math.clz32(i);
 i = Math.imul(i, i);
